Incorporates the CMSIS DSP library and demo wav
I added the CMSIS DSP library, which was painless... mostly. I haven't had a lot of luck with audio playback though.
The output from the DAC sounds like a buzz saw.

There's now a raw unsigned 8-bit 11025Hz WAV file programmed to flash. This is copied to RAM in main.c. From there,
the data is pushed to the DAC by the DMA.

The audio file was created with Audacity and xxd. The relevant command was

xxd -i guitar.raw guitar.h